Refactor from node targeting to era targeting. Also makes the era ENV extendable
