+++ source/environment.ts
@@ -1,10 +1,10 @@
 const isNavigatorDefind = typeof navigator !== 'undefined';
 
 export const environment = {
+  CARDANO_ERA: process.env.CARDANO_ERA,
   DEBUG: process.env.DEBUG,
   GRAPHQL_HTTP_URL: `${process.env.GRAPHQL_API_PROTOCOL}://${process.env.GRAPHQL_API_HOST}:${process.env.GRAPHQL_PORT}/`,
   GRAPHQL_WEBSOCKET_URL: `${process.env.GRAPHQL_WEBSOCKET_PROTOCOL}://${process.env.GRAPHQL_WEBSOCKET_HOST}:${process.env.GRAPHQL_PORT}/`,
   IS_CLIENT: isNavigatorDefind,
-  IS_RUST_CLIENT: process.env.IS_RUST_CLIENT,
   IS_SERVER: !isNavigatorDefind,
 };
